requirements.txt
asgiref==3.11.1
bleach==6.3.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
cryptography==46.0.7
dj-database-url==3.1.2
Django==5.2.13
django-allauth==65.16.0
django-htmx==1.27.0
greenlet==3.4.0
gunicorn==25.3.0
idna==3.11
jwt==1.4.0
packaging==26.1
playwright==1.58.0
psycopg2-binary==2.9.11
pycparser==3.0
pyee==13.0.1
python-decouple==3.8
python-slugify==8.0.4
requests==2.33.1
sqlparse==0.5.5
text-unidecode==1.3
typing_extensions==4.15.0
urllib3==2.6.3
webencodings==0.5.1
whitenoise==6.12.0
